requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
google-generativeai>=0.3.1
vaderSentiment>=3.3.2
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
python-dotenv>=1.0.0
reportlab>=4.0.0
Pillow>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0

